Refactor backend to backend/javascript
