+++ src/backend/javascript-helpers.js
